requirements.txt
google-adk==1.13.0
ddgs==9.5.4